refactor(runtime): extract main lifecycle responsibilities into modules
